+++ apis/Ids.h
@@ -337,42 +337,42 @@ namespace Exchange {
 
         ID_FIRMWARE_DOWNLOAD                         = ID_ENTOS_OFFSET + 0x460,
         ID_FIRMWARE_DOWNLOAD_NOTIFICATION            = ID_FIRMWARE_DOWNLOAD + 1,
-        
-        ID_DEVICESETTINGS_AUDIO                             = ID_ENTOS_OFFSET + 0x470,
-        ID_DEVICESETTINGS_AUDIO_NOTIFICATION                = ID_DEVICESETTINGS_AUDIO + 1,
-        ID_DEVICESETTINGS_AUDIO_ENCODING_ITERATOR           = ID_DEVICESETTINGS_AUDIO + 2,
-        ID_DEVICESETTINGS_AUDIO_COMPRESSION_ITERATOR        = ID_DEVICESETTINGS_AUDIO + 3,
-        ID_DEVICESETTINGS_AUDIO_PROFILE_ITERATOR            = ID_DEVICESETTINGS_AUDIO + 4,
-        ID_DEVICESETTINGS_AUDIO_STEREOMODE_ITERATOR         = ID_DEVICESETTINGS_AUDIO + 5,
 
-        ID_DEVICESETTINGS_COMPOSITEIN                       = ID_ENTOS_OFFSET + 0x480,
-        ID_DEVICESETTINGS_COMPOSITEIN_NOTIFICATION          = ID_DEVICESETTINGS_COMPOSITEIN + 1,
+        ID_USER_PLUGIN                               = ID_ENTOS_OFFSET + 0x470,
 
-        ID_DEVICESETTINGS_DISPLAY                           = ID_ENTOS_OFFSET + 0x490,
-        ID_DEVICESETTINGS_DISPLAY_NOTIFICATION              = ID_DEVICESETTINGS_DISPLAY + 1,
-        ID_DEVICESETTINGS_DISPLAY_RESOLUTION_ITERATOR       = ID_DEVICESETTINGS_DISPLAY + 2,
-        ID_DEVICESETTINGS_DISPLAY_HOTPLUG_NOTIFICATION      = ID_DEVICESETTINGS_DISPLAY + 3,
+        ID_DEVICESETTINGS_MANAGER_AUDIO                             = ID_ENTOS_OFFSET + 0x480,
+        ID_DEVICESETTINGS_MANAGER_AUDIO_NOTIFICATION                = ID_DEVICESETTINGS_MANAGER_AUDIO + 1,
+        ID_DEVICESETTINGS_MANAGER_AUDIO_ENCODING_ITERATOR           = ID_DEVICESETTINGS_MANAGER_AUDIO + 2,
+        ID_DEVICESETTINGS_MANAGER_AUDIO_COMPRESSION_ITERATOR        = ID_DEVICESETTINGS_MANAGER_AUDIO + 3,
+        ID_DEVICESETTINGS_MANAGER_AUDIO_PROFILE_ITERATOR            = ID_DEVICESETTINGS_MANAGER_AUDIO + 4,
+        ID_DEVICESETTINGS_MANAGER_AUDIO_STEREOMODE_ITERATOR         = ID_DEVICESETTINGS_MANAGER_AUDIO + 5,
 
-        ID_DEVICESETTINGS_FPD                               = ID_ENTOS_OFFSET + 0x4A0,
-        ID_DEVICESETTINGS_FPD_NOTIFICATION                  = ID_DEVICESETTINGS_FPD + 1,
- 
-        ID_DEVICESETTINGS_HDMIIN                            = ID_ENTOS_OFFSET + 0x4B0,
-        ID_DEVICESETTINGS_HDMIIN_NOTIFICATION               = ID_DEVICESETTINGS_HDMIIN + 1,
-        ID_DEVICESETTINGS_HDMIIN_PORTCONNECTION_ITERATOR    = ID_DEVICESETTINGS_HDMIIN + 2,
-        ID_DEVICESETTINGS_HDMIIN_GAMELIST_ITERATOR          = ID_DEVICESETTINGS_HDMIIN + 3,
+        ID_DEVICESETTINGS_MANAGER_COMPOSITEIN                       = ID_ENTOS_OFFSET + 0x490,
+        ID_DEVICESETTINGS_MANAGER_COMPOSITEIN_NOTIFICATION          = ID_DEVICESETTINGS_MANAGER_COMPOSITEIN + 1,
 
-        ID_DEVICESETTINGS_HOST                              = ID_ENTOS_OFFSET + 0x4C0,
-        ID_DEVICESETTINGS_HOST_NOTIFICATION                 = ID_DEVICESETTINGS_HOST + 1,
+        ID_DEVICESETTINGS_MANAGER_DISPLAY                           = ID_ENTOS_OFFSET + 0x500,
+        ID_DEVICESETTINGS_MANAGER_DISPLAY_NOTIFICATION              = ID_DEVICESETTINGS_MANAGER_DISPLAY + 1,
+        ID_DEVICESETTINGS_MANAGER_DISPLAY_RESOLUTION_ITERATOR       = ID_DEVICESETTINGS_MANAGER_DISPLAY + 2,
+        ID_DEVICESETTINGS_MANAGER_DISPLAY_HOTPLUG_NOTIFICATION      = ID_DEVICESETTINGS_MANAGER_DISPLAY + 3,
 
-        ID_DEVICESETTINGS_VIDEODEVICE                       = ID_ENTOS_OFFSET + 0x4D0,
-        ID_DEVICESETTINGS_VIDEODEVICE_NOTIFICATION          = ID_DEVICESETTINGS_VIDEODEVICE + 1,
-        ID_DEVICESETTINGS_VIDEO_CODEC_PROFILE_ITERATOR      = ID_DEVICESETTINGS_VIDEODEVICE + 2,
+        ID_DEVICESETTINGS_MANAGER_FPD                               = ID_ENTOS_OFFSET + 0x510,
+        ID_DEVICESETTINGS_MANAGER_FPD_NOTIFICATION                  = ID_DEVICESETTINGS_MANAGER_FPD + 1,
 
-        ID_DEVICESETTINGS_VIDEOPORT                         = ID_ENTOS_OFFSET + 0x4E0,
-        ID_DEVICESETTINGS_VIDEOPORT_NOTIFICATION            = ID_DEVICESETTINGS_VIDEOPORT + 1,
-        
-        //Test plugin for Devicesettings plugin. TODO: will remove in final review PR
-        ID_USER_PLUGIN                               = ID_ENTOS_OFFSET + 0x4F0
+        ID_DEVICESETTINGS_MANAGER_HDMIIN                            = ID_ENTOS_OFFSET + 0x520,
+        ID_DEVICESETTINGS_MANAGER_HDMIIN_NOTIFICATION               = ID_DEVICESETTINGS_MANAGER_HDMIIN + 1,
+        ID_DEVICESETTINGS_MANAGER_HDMIIN_PORTCONNECTION_ITERATOR    = ID_DEVICESETTINGS_MANAGER_HDMIIN + 2,
+        ID_DEVICESETTINGS_MANAGER_HDMIIN_GAMELIST_ITERATOR          = ID_DEVICESETTINGS_MANAGER_HDMIIN + 3,
+
+        ID_DEVICESETTINGS_MANAGER_HOST                              = ID_ENTOS_OFFSET + 0x530,
+        ID_DEVICESETTINGS_MANAGER_HOST_NOTIFICATION                 = ID_DEVICESETTINGS_MANAGER_HOST + 1,
+
+        ID_DEVICESETTINGS_MANAGER_VIDEODEVICE                       = ID_ENTOS_OFFSET + 0x540,
+        ID_DEVICESETTINGS_MANAGER_VIDEODEVICE_NOTIFICATION          = ID_DEVICESETTINGS_MANAGER_VIDEODEVICE + 1,
+        ID_DEVICESETTINGS_MANAGER_VIDEO_CODEC_PROFILE_ITERATOR      = ID_DEVICESETTINGS_MANAGER_VIDEODEVICE + 2,
+
+        ID_DEVICESETTINGS_MANAGER_VIDEOPORT                         = ID_ENTOS_OFFSET + 0x550,
+        ID_DEVICESETTINGS_MANAGER_VIDEOPORT_NOTIFICATION            = ID_DEVICESETTINGS_MANAGER_VIDEOPORT + 1
+
     }; // enum IDS
 
 } // namespace Exchange
